third_party/WebKit/LayoutTests/http/tests/serviceworker/chromium/xhr-is-not-exposed-to-service-workers-expected.txt

This is a testharness.js-based test.
Harness Error. harness_status.status = 1 , harness_status.message = 1 duplicate test name: "XMLHttpRequest is not exposed to ServiceWorkerGlobalScope"
PASS XMLHttpRequest is not exposed to ServiceWorkerGlobalScope
PASS XMLHttpRequest is not exposed to ServiceWorkerGlobalScope
Harness: the test ran to completion.
